jsanchez-WORK. Finish deploy MedicalExaminationDocuments

+++ Solution/Prevea.Service/Service/ServiceMedicalExaminationDocuments.cs
@@ -6,6 +6,9 @@
     using IService.IService;
     using Model.Model;
     using System;
+    using System.Web;
+    using System.Diagnostics;
+    using System.IO;
 
     #endregion
 
@@ -25,8 +28,42 @@
         {
             try
             {
-                medicalExaminationDocument = Repository.SaveMedicalExaminationDocument(medicalExaminationDocument);
+                if (string.IsNullOrEmpty(medicalExaminationDocument.Enrollment))
+                {
+                    string enrollment;
+                    var countMaxDocuments = Repository.GetMaxMedicalExaminationDocumentByType(
+                        medicalExaminationDocument.RequestMedicalExaminationEmployeeId,
+                        medicalExaminationDocument.MedicalExaminationDocumentTypeId) + 1;
 
+                    switch (medicalExaminationDocument.MedicalExaminationDocumentTypeId)
+                    {
+                        case (int)EnMedicalExaminationDocumentType.BloodTest:
+                            enrollment = $"ME_BT.{medicalExaminationDocument.RequestMedicalExaminationEmployeeId}.{countMaxDocuments}";
+                            break;
+                        case (int)EnMedicalExaminationDocumentType.Electrocardiogram:
+                            enrollment = $"ME_EL.{medicalExaminationDocument.RequestMedicalExaminationEmployeeId}.{countMaxDocuments}";
+                            break;
+                        case (int)EnMedicalExaminationDocumentType.AudiometricReport:
+                            enrollment = $"ME_AR.{medicalExaminationDocument.RequestMedicalExaminationEmployeeId}.{countMaxDocuments}";
+                            break;
+                        case (int)EnMedicalExaminationDocumentType.Spirometry:
+                            enrollment = $"ME_SP.{medicalExaminationDocument.RequestMedicalExaminationEmployeeId}.{countMaxDocuments}";
+                            break;
+                        case (int)EnMedicalExaminationDocumentType.UrineAnalytics:
+                            enrollment = $"ME_UA.{medicalExaminationDocument.RequestMedicalExaminationEmployeeId}.{countMaxDocuments}";
+                            break;
+                        case (int)EnMedicalExaminationDocumentType.Others:
+                            enrollment = $"ME_OT.{medicalExaminationDocument.RequestMedicalExaminationEmployeeId}.{countMaxDocuments}";
+                            break;
+
+                        default:
+                            enrollment = $"ME_OT.{medicalExaminationDocument.RequestMedicalExaminationEmployeeId}.{countMaxDocuments}";
+                            break;
+                    }
+                    medicalExaminationDocument.Enrollment = enrollment;
+                }
+
+                medicalExaminationDocument = Repository.SaveMedicalExaminationDocument(medicalExaminationDocument);
                 if (medicalExaminationDocument == null)
                 {
                     return new Result
@@ -58,10 +95,22 @@
 
         public Result DeleteMedicalExaminationDocument(int id)
         {
-                        try
+            try
             {
-                var result = Repository.DeleteMedicalExaminationDocument(id);
+                var medicalExaminationDocument = Repository.GetMedicalExaminationDocumentById(id);
+                var physicalPath = HttpContext.Current.Server.MapPath(medicalExaminationDocument.Url);
+                if (!RemoveFile(physicalPath))
+                {
+                    return new Result
+                    {
+                        Message = "Se ha producido un error al Borrar el Documento",
+                        Object = null,
+                        Status = Status.Error
+                    };
 
+                }
+
+                var result = Repository.DeleteMedicalExaminationDocument(id);
                 if (result == false)
                 {
                     return new Result
@@ -70,7 +119,7 @@
                         Object = null,
                         Status = Status.Error
                     };
-                }
+                }                
 
                 return new Result
                 {
@@ -94,6 +143,42 @@
         {
             return Repository.GetMedicalExaminationDocumentsByRequestMedicalExaminationEmployeeId(
                 requestMedicalExaminationEmployeeId);
+        }
+
+        public Result SaveFileMedicalExaminationDocument(HttpPostedFileBase fileDocument, int medicalExaminationDocumentId)
+        {
+            try
+            {
+                var medicalExaminationDocument = GetMedicalExaminationDocumentById(medicalExaminationDocumentId);
+                if (medicalExaminationDocument == null)
+                    return new Result { Status = Status.Error };
+
+                var requestMedicalExaminationEmployee =
+                    Repository.GetRequestMedicalExaminationEmployeeById(medicalExaminationDocument
+                        .RequestMedicalExaminationEmployeeId);
+
+                medicalExaminationDocument.Url = $"~/App_Data/Companies/{requestMedicalExaminationEmployee.RequestMedicalExaminations.Company.NIF}/MedicalExaminations/{medicalExaminationDocument.Enrollment}.pdf";
+
+                var url = Path.Combine(HttpContext.Current.Server.MapPath(medicalExaminationDocument.Url));
+                fileDocument.SaveAs(url);
+
+                var result = SaveMedicalExaminationDocument(medicalExaminationDocument);
+                if (result.Status == Status.Error)
+                    return new Result { Status = Status.Error };
+
+                return new Result
+                {
+                    Status = Status.Ok,
+                    Object = result.Object
+                };
+            }
+            catch (Exception ex)
+            {
+                Debug.WriteLine(ex.Message);
+
+                return new Result { Status = Status.Error };
+            }
+
         }
     }
 }
